Summary: minor, always pass on SXE_DYLD_PATH with pointers to the lwlibdir Keywords: bldchn
This changeset prepends calls to the undumped or dumped sxe binary with
BATCHENV which holds paths to the newly built lwlib libs.  This fixes
builds that switch from say the athena widget kit to the lucid kit, as
the installed lwlib library will certainly not provide some symbols
needed with the new widget kit.

* lisp/Makefile.am: prepend all calls to sxe with BATCHENV

* modules/Makefile.am: prepend all calls to sxe with BATCHENV

* src/Makefile.am: prepend all calls to sxe with BATCHENV

Signed-off-by: Sebastian Freundt <freundt@ga-group.nl>