Summary: fix, actually dispatch events, fixes bug #131 Keywords: bug fix #131
This changeset makes sure that the communication between REPL event
queue and delegation event queue (the one for the worker threads)
actually work in both directions.  Events put on the REPL event queue
from within jobs that have been delegated are now announced to the event
queue handler and hence executed, which in particular was the reason the
sentinel wasn't invoked as described in bug #131.

* src/events/worker-asyneq.c: call asyneq_handle_event() after enqueuing

Signed-off-by: Sebastian Freundt <freundt@ga-group.nl>