Summary: minor, provide malldbg multiarg for MALLOC_PERTURB_ Keywords: bldchn diagnostics
This changeset introduces a multiarg item to the error-checking option
group, named `malldbg', that controls MALLOC_PERTURB_ and friends when
dumping.  This should fix the dumping procedure on MALLOC_PERTURB_
systems as our dumper reacts quite sensitively to that kind of system
hardening.

This changeset contains no additional intelligence to turn it off
automatically, if you see crashes during the dump phase conside
--with-error-checking=nomalldbg.

Whether or not we can persuade (read fix) the dumper code to react more
friendly to this kind of interference is subject to a different patch.

* configure.ac: add `malldbg' to the error-checking multiarg

* m4/sxe-compiler.m4: prepare EXTRA_BATCHENV in case malldbg is off

* src/Makefile.am: extend dumping environment to EXTRA_BATCHENV

Signed-off-by: Sebastian Freundt <hroptatyr@fresse.org>