Fix the age old problem of TTY cursor being reset on exit
This changeset stops SXEmacs from resetting the text cursor to block on
exit on a Linux console.  It is actually still reset, but it is reset
correctly and now you'll end up with a "normal" cursor instead of a "very
visible" one.

One side-effect of this change is that SXEmacs on a console will now use a
block cursor, but that just adds consistency because that is how things
are in X11 and Xterm anyway.

	* src/ui/TTY/console-tty.h (struct tty_console): Add
	`cursor_invisible_' and `hard_cursor_' for invisible and hard to
	see cursors.

	* src/ui/TTY/redisplay-tty.c (init_tty_for_redisplay): Set
	cursor_visible_ to "vs" and cursor_normal_ to "ve" like they
	should be according to terminfo(5).
	Also add cursor_invisible_ ("vi"), and hard_cursor_ ("HC").

Signed-off-by: Steve Youngs <steve@sxemacs.org>